Fix SiftGPU bug with skipped levels. (#396)
If a image level is skipped due to a limit on the number of features, this level's feature counter can still contain data from the last image. This can lead to a heap corruption when downloading the keypoints from GPU. It might also cause reusing keypoint locations from the last image (not verified).