Backwards compatibility of model_aligner (#1240)
In previous versions of Colmap, the `model_aligner` would align a 3D model to a set of camera poses. In this version, not setting the `alignment_type` option defaults to aligning the model with the dominant plane, which is not the behavior users would expect given previous versions (see #1239 ). I suggest to keep backwards compatibility by keeping the original behavior as the default.