Fix rare deadlock during matching due to concurrent database access (#2876)
It turns out one cannot concurrent read from the database and begin a
transaction. This could happen in rare cases during vocabulary tree
based matching, as the querying of the vocabulary tree is performed
concurrently to the main matching controller thread. In this case, one
thread might simultaneously read and begin a transaction. This causes a
deadlock if unlucky. With this PR, this cannot happen anymore.