Re-enable interface support for covariance estimation from a Ceres::Problem instance (#3061)
Sincere request to re-enable this feature. Exposing this interface will
make dev much easier when problem is constructed with pyceres, for
example, with ``pycolmap.cost_functions.ReprojErrorCost`` and
non-identity keypoint covariance (e.g. with a scaled covariance due to
resizing). Given that we expose the ``pycolmap.cost_functions`` module
there is no downside to support this.