Fix covariance propagation of pose inverse (#3155)
The covariance propagation of pose inverse via adjoint introduced in
https://github.com/colmap/colmap/pull/2598 followed the lamar
convention, which was just found to be actually broken here:
https://github.com/microsoft/lamar-benchmark/pull/84.

This PR fixes the propagation and added more cycle-consistent tests on
the property of the adjoint matrix. To simplify the code for calculating
adjoint inverse the function ``CrossProductMatrix`` is moved into
``rigid3.h``.