Add docs on the left convention in COLMAP for covariance propagation. (#3167)
This was a big confusion. I added some comments to explain why the fix
in https://github.com/colmap/colmap/pull/3155 was correct.

In the end its because that in COLMAP we use the left convention, while
in GTSAM and its reference paper here https://arxiv.org/pdf/1812.01537
the right convention was employed.

This is also in preparation for another PR on propagating relative pose
covariance in COLMAP (with left convention) with cross-pose correlation
available. This will encode the fact that the close image has lower
relative pose covariance and does not get affected by Gauge ambiguity.