Update FindMetis.cmake to also link GK_LIBRARIES (#3470)
I might be missing something, but I was encountering missing symbols
from GKlib during the linking stage. It seems like an oversight:
FindMetis.cmake correctly locates the GKlib library, but it doesn't
actually link it after finding it. This pull request adds GKlib to the
list of linked libraries to resolve that issue.

---------

Co-authored-by: Johannes Schönberger <jsch@demuc.de>