Avoid frame having data from a sensor that does not exist in its associated rig.  (#3543)
As a follow up to https://github.com/colmap/colmap/pull/3540 for safety
improvement, this PR includes the following fixes:

* ``AddFrame`` should not add a frame with empty data as its never
useful in the reconstruction. NumData() is added for the frame class.
* ``Frame.SetRigPtr`` should set a rig that has all sensors in the frame
data supported.
* Conversely, when a rig pointer is available for the frame,
``Frame.AddDataId`` should add data that with the sensor available in
the associated rig.

Related tests are added.

---------

Co-authored-by: Johannes Schönberger <jsch@demuc.de>