Add bindings for mvs::model for covisibility support in pycolmap (#3621)
A step to make covisibility support available in pycolmap, for easier
pair generation. For example, this could speed up the python for loop
from hloc here:
https://github.com/cvg/Hierarchical-Localization/blob/master/hloc/pairs_from_covisibility.py#L16-L47
and also simplify the code from the python side.

Ideally it might be better to move the whole covisibility logic to the
reconstruction utility, but this will require adding some code to
convert PMVS to colmap format, so is only left to future work for now.