Fix broken logic of removing cameras at Reconstruction::TearDown(). (#3634)
With the new abstraction of rigs and frames, at
``Reconstruction::TearDown()``, the removal of cameras should correspond
to the removal of rigs. Otherwise the reconstruction object will be
broken after TearDown().

In practice, This PR only fixes corner cases when no associated
images are available for a camera in the rig.