Also replace mask file extension to .png instead of appending (#3611)
Sometimes the input images are a mix of .png and other extensions.
To handle the mask filenames, which are always .png, it is cleaner to
replace the image extension instead of appending.
This would avoid filenames such as **filename.jpg.png** and
**filename.png.png**

Change:
If checking fails for appending .png to the mask path, I also check by
replacing the file extension by .png.

---------

Co-authored-by: Johannes Schönberger <jsch@demuc.de>