Adjust reconstruction consistency checks (#3658)
It's hugely inconvenient to require data to already be in the frame when
a frame is added to the reconstruction and it broke a bunch of internal
code. Instead, I am tightening some of the other checks, where I am now
checking that when adding an image, the corresponding frame has the
image and the equivalent for the sensors in frames.