Fix database race conditions in mapping pipelines (#3865)
Opening a database is not thread-safe due to the incurred database
schema migration and PRAGMA operations, which modify the database. In
particular, the concurrent incremental mappers in the hierarchical
mapper currently sometimes lead to errors when opening the database
access at the same time with an error "database is locked". This then
silently fails the reconstruction of the leaf node cluster (as the
exception is swallowed by the thread pool, which will be addressed in a
separate PR).

For consistency across all mapping pipelines, we now take the database
object to construct the pipelines. Likewise, the image_path is now
consistently specified via the pipeline options.