Add rotation averaging controller in colmap and consolidate glomap exe functions. (#3888)
Final PR of the rotation averaging update serie. 

This PR includes:
* Add rotation averaging controller in glomap, which takes in a database
and runs view graph calibration, relative pose estimation, and rotation
averaging. Add nominal test.
* Move rotation averaging function from sfm to estimators. 
* Consolidate the two functions in glomap/exe into glomap.h/cc, in
preparation for being moved to colmap/exe.

The glomap rotation averaging test is kept in the sfm folder, because it
actually uses global mapper and skips the steps after rotation
averaging. To be able to really achieve a stand-alone case of rotation
averaging, we would need to handle the IO. For now this is done by
reading in the files with glomap/pose_io.h but in the future this can be
done better by writing a solid I/O for ViewGraph (or probably renamed
PoseGraph in the future and live in colmap).