Simplify the logic and naming of reconstruction pruning.  (#3886)
After all skip_pruning is by default set to true. I wonder when this
block of code was tested last time.

The logic change of this PR should be equivalent. The follow up would be
to move both "reconstruction pruning" and "view graph manipulation" to
sth like "global_mapper_impl".